# Switch node authorizer indexes to reference counts, add fastpath edge removal