**What type of PR is this?**
/kind cleanup

**What this PR does / why we need it**:
Improves node authorizer index maintenance by switching to reference counts and adding a fastpath for edge removal.

Performance compared to current master:
```
benchmark                            old ns/op      new ns/op      delta
BenchmarkWriteIndexMaintenance-12    875808         8325           -99.05%

benchmark                            old allocs     new allocs     delta
BenchmarkWriteIndexMaintenance-12    41             38             -7.32%

benchmark                            old bytes      new bytes      delta
BenchmarkWriteIndexMaintenance-12    1924           1601           -16.79%
```

**Does this PR introduce a user-facing change?**:
```release-note
Improves performance of the node authorizer
```

/cc @mikedanese @wojtek-t @mborsz